requirements.txt
Django>=3.2,<=4
# psycopg2>=2.9.3,<2.10
djangorestframework==3.13.1
drf-spectacular>=0.23.1,<0.24
django-cors-headers>=3.13, <3.14
